refactor(tests): consolidate media generator fixtures into conftest
